refactor(recruitment): 지원 커맨드 트랜잭션을 RecruitmentApplicationProcessor로 이관
- RecruitmentApplicationService의 지원 생성/상태변경/취소/메시지 로직을
  @Component @Transactional RecruitmentApplicationProcessor로 이관
- Processor가 자기 트랜잭션 안에서 엔티티를 ID로 로드(load-in-tx)하여
  detached 더티체킹 유실/LazyInit 방지, 상태변경+그룹 편입 원자성 보장
- Service는 Processor 위임만 담당하도록 축소(§5.4)
- 오케스트레이션 테스트를 RecruitmentApplicationProcessorTest로 이관,
  ServiceTest는 위임 검증으로 전환
